+++ eNMS/models/scheduling.py
@@ -1,4 +1,3 @@
-from apscheduler.triggers.cron import CronTrigger
 from datetime import datetime
 from re import search
 from sqlalchemy import Boolean, case, ForeignKey, Integer
@@ -81,14 +80,6 @@ class Task(AbstractBase):
             return f"{days}{hours}h:{minutes}m:{seconds}s"
         return None
 
-    def aps_conversion(self, date):
-        dt: datetime = datetime.strptime(date, "%d/%m/%Y %H:%M:%S")
-        return datetime.strftime(dt, "%Y-%m-%d %H:%M:%S")
-
-    def aps_date(self, datetype):
-        date = getattr(self, datetype)
-        return self.aps_conversion(date) if date else None
-
     def pause(self):
         self.is_active = False
         db.session.commit()
@@ -108,55 +99,13 @@ class Task(AbstractBase):
             properties["pools"] = [pool.id for pool in self.pools]
         return properties
 
-    def kwargs(self):
-        default = {
-            "id": self.aps_job_id,
-            "func": app.run,
-            "replace_existing": True,
-            "args": [self.service.id],
-            "kwargs": self.run_properties(),
-        }
-        if self.scheduling_mode == "cron":
-            self.periodic = True
-            expression = self.crontab_expression.split()
-            mapping = {
-                "0": "sun",
-                "1": "mon",
-                "2": "tue",
-                "3": "wed",
-                "4": "thu",
-                "5": "fri",
-                "6": "sat",
-                "7": "sun",
-                "*": "*",
-            }
-            expression[-1] = ",".join(mapping[day] for day in expression[-1].split(","))
-            trigger = {"trigger": CronTrigger.from_crontab(" ".join(expression))}
-        elif self.frequency:
-            self.periodic = True
-            frequency_in_seconds = (
-                int(self.frequency)
-                * {"seconds": 1, "minutes": 60, "hours": 3600, "days": 86400}[
-                    self.frequency_unit
-                ]
-            )
-            trigger = {
-                "trigger": "interval",
-                "start_date": self.aps_date("start_date"),
-                "end_date": self.aps_date("end_date"),
-                "seconds": frequency_in_seconds,
-            }
-        else:
-            self.periodic = False
-            trigger = {"trigger": "date", "run_date": self.aps_date("start_date")}
-        return default, trigger
-
     def schedule(self):
-        default, trigger = self.kwargs()
         if not app.scheduler.get_job(self.aps_job_id):
             app.scheduler.add_job(**{**default, **trigger})
         else:
             app.scheduler.reschedule_job(default.pop("id"), **trigger)
+
+
 
 
 @db.set_custom_properties
